v2: corrective retrieval agent, observability, τρία νέα golden sets
RETRIEVAL
- corrective retry μετά από κοπή του gate: rewrite + 2ο pass με αυστηρότερο
  κατώφλι (+1.0). 4 ερωτήσεις από σιωπή σε σωστή απάντηση, 0 ψευδαισθήσεις,
  out_of_corpus 5/5, κύριο σετ 61/61 ανέπαφο. Μηδέν κόστος στο happy path.
- domain-aware translation prompt: «μηχανήματα»→machinery→provisioning,
  «ελέγχου»→control→auditability. Έλεγχος: 30 αγγλικές, max |Δ| = 0.000.

OBSERVABILITY
- metrics.py + /metrics (Prometheus text, 0 εξαρτήσεις, 0 νέα containers).
  Νέα ορατότητα: gate_block_rate, corrective_success_rate, tokens, latency.

EVALUATION
- golden_multihop_new (11 cross-document), golden_hard_paraphrase (16 stress
  test), golden_conversations (12 πολύγυρες + 2 leak tests).
- διορθώσεις μέτρησης q047/q044: keywords που δεν υπήρχαν στο corpus.
  baseline 0.747/97.0% -> 0.770/98.5% (ΟΧΙ όλο βελτίωση — δύο κρυμμένες διορθώσεις μέτρησης).
- πρώτη μέτρηση conversational rewriting: 40% -> 90% (oracle 100%).

ΑΠΟΡΡΙΦΘΗΚΑΝ με μέτρηση
- query decomposition: 3 στρατηγικές, καλύτερη +1 keyword σε 45 για +50%
  latency σε κάθε ερώτηση.
- λακωνικό corrective prompt: χειρότερο σε κάθε κατώφλι, και καταστρέφει τη
  δυνατότητα φιλτραρίσματος.

TESTS: 56 -> 72 (test_corrective_retrieval, test_metrics). CI ενημερωμένο.

+++ backend/loadtest.py
@@ -30,7 +30,7 @@ def _pct(values, p):
     if not values:
         return float("nan")
     s = sorted(values)
-    k = max(0, min(len(s) - 1, int(round((p / 100) * (len(s) - 1)))))
+    k = max(0, min(len(s) - 1, round((p / 100) * (len(s) - 1))))
     return s[k]
